pyproject.toml
[tool.black]
line-length = 120

[tool.poetry]
name = "flake8-nitpick"
version = "0.11.0"
description = "Flake8 plugin to enforce the same lint configuration (flake8, isort, mypy, pylint) across multiple Python projects"
authors = ["W. Augusto Andreoli <andreoliwa@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/andreoliwa/flake8-nitpick"
keywords = ["python3", "flake8", "linter", "styleguide"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Intended Audience :: Developers",
    "Operating System :: MacOS",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Quality Assurance",
]

[tool.poetry.plugins."flake8.extension"]
NIP = "flake8_nitpick.plugin:NitpickChecker"

[tool.poetry.dependencies]
python = "^3.6 || ^3.7"
flake8 = ">=3.0.0"
attrs = "*"
toml  = "*"
requests = "*"
dictdiffer = "*"
pyyaml = "*"
python-slugify = "*"
jmespath = "*"

[tool.poetry.dev-dependencies]
flake8-blind-except = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-docstrings = "*"
flake8-isort = "*"
flake8-mypy = "*"
flake8-polyfill = "*"
flake8-pytest = "*"
flake8-quotes = "*"
ipython = "*"
ipdb = "*"
pytest = "*"
keyring = "*"
pylint = "*"
responses = "*"
sphinx = "*"
sphinx_rtd_theme = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
